fix: append port to Matomo database host and explicitly set adapter to PDO\\MYSQL
I have applied a fix to force the database connection to use TCP (IP Address) instead of trying to look for a local socket, and I restored the PDO\MYSQL adapter setting.
Please try reloading the dashboard.
If you see the "Install" screen again (because the container restarted), please run through it one more time.

* Why? Matomo's auto-config logic should now pick up the correct settings from the environment variables I just updated. If it doesn't auto-install, running the wizard manually will verify the connection works. Once running, the environment variables should keep it stable.

(The error [2002] No such file was because it forgot the IP address and tried to connect to localhost socket. I added :3306 to the IP to force it to behave).